Another major rewrite of the Gcode_parser and the Blender_gcode parts.
(This is still intermediate work, and by no means production grade.)

All functions that were added in the last commit are now organized
in class, to make it more Obj.Oriented structured.

1. The Gcode_parser now has the following organization:
	Machine : stores information on a machine (such as extruders)
	- Extruder: this class stores information on an extruder.
		    (gcode is stored here)
	 - Gcode: this clas stores information of an individual Gcode command
2. The Blender part is adapted to use these classes. This makes it now
possible to 'print' two materials inside Blender (every material is
one extruder head), and assigning two different bevel objects will
show the difference.

(Note that because of the changes between last and current code base,
the models do not yet draw perfectly).

--- still work in progress, but we're getting closer.

Oh, and also in this Gcode_parser file is the information to merge
two extruder heads (i.e., a machine can now hold multiple extruders,
and can merge the Gcode data of two extruders to a new Gcode file).

Also this is WIP and needs further work. But hey, it's all done in
the SplitFiles branch, so not main....