The Gcode class instances were initiated with a set X,Y, and Z (X=0.0, Y=0.0, Z=0.0). However, this resulted in an issue with updating the Gcode command with the Gcode.update_state() method when only the E or F values were modified. In that case, the X,Y,and Z were not None and thus the lastState values were updated with X=0.0, Y=0.0, and Z=0.0 values where they shouldn't be.
With this commit, the X, Y, and Z values are set to None initially, which
is physically more accurate (as no information is available on its position).

This might break things at some locations (not sure if there is a test
for the NoneType for X, Y, or Z) but as this is the closest to
physical reality, I'd advocate to leave it like this and modify any
code that has issues with this.