Modified the 'debug_extruder' functionality once more.
This commit fixes an issue when debugging an extruder that contains
merged data from multiple extruders: this extruder does not have
any 'rawGcode' commands, and thus the output file contained nothing.

Solved by checking if the length of rawGcode > 0. If yes, add the
rawGcode command; if no, ignore and just print the standardGcode.