Include the functionality to import .gcode file and to show the curves inside Blender, without the need to do so by loading the 'Blender_gcode.py' file into the Blender Script Editor.
With this __init__.py file, now a user is able to select the addon
from the user preferences (Addon >> Import/Export/). When activated,
both a file menu entry appears as well as a panel (where a check
can be placed to use a bevel object or not). When the file is loaded,
a blender GcurveData object is drawn onto the screen.

(This is similar as running the Blender_gcode.py file from a script)

This code works with the latest revision as of April 21, see version info
for the exact blender revision nr.