Include the 'Bottle' micro webframework
The Bottle web framework (see www.bottlepy.org) is a small, single
file webframework that is able to run a local server (and lots of
other things).

Given that we need an interface for the Gcode parser, and with
many different operating systems that might use this, I figured
it might be just as easy to communicate with the Gcode parser
through HTML.

The 'LICENCE' file is renamed to 'bottle_LICENCE' to indicate
that that license file only deals with the 'bottle.py' file
and not with the other files.