A working, initial web interface to merge Gcode
Put in place the files that are required to provide a web
interface using the 'bottle' microframework. The html files
'index' and 'merge_gcode'.html contain the web pages in which
information is entered; the /webInterface files and directories
store all the further information that is needed.

Static --> directory that holds static files, such as images and css
stylesheet
Templates --> directory that holds template files, if any
Tmp --> directory in which temporarily the files to be merged
are stored in
.webInterface --> the base file, this folder contains the .html files

The script 'gcode_webinterface.py' can be run from the command line,
using 'python3 gcode_webinterface.py'; it should start the Bottle
server, and in a browser on 'localhost:8080' there should be the
web interface up and running.