The merging of two .gcode files with webpage is functional.
The modifications in 'gcode_webinterface.py' ensure that now
a webpage is loaded with entries for two .gcode files. After
a basic and simple check for validity (must be improved!),
the files are read and and 'Ultimaker virtual machine' is
created. The files are loaded in their respective virtual
'extruders' and processed. The result is a merged file, which
is saved on disk with a default name ('merged.gcode').

Changes in the .css and .html files are of cosmetic form only,
it's moving the button and label fields a bit, and make the page
a bit wider so I don't have to scroll.